// src/domain.types/reservation.order.item/reservation.order.item.dto.ts
/* eslint-disable linebreak-style */
/* eslint-disable eol-last */
/* eslint-disable linebreak-style */

export interface ReservationOrderItemDto {
    id?: string;
    ReservationOrderId: string;
    RoomId: string;
    Tax: string;
    Discount: string;
    Cost: string;
    TotalDays:string;
    
}